weave in backtest time

+++ handler.py
@@ -19,10 +19,10 @@ def write_records(client, current_value, algorithm, env, portfolio, exchange, da
     else:
         try:
             date = datetime.strptime(backtest_time, '%Y-%m-%d %H:%M:%S.%f')
-            current_time = str(time.mktime(date.timetuple()))
+            current_time = str(int(time.mktime(date.timetuple())))
         except Exception:
             date = datetime.strptime(backtest_time, '%Y-%m-%d %H:%M:%S')
-            current_time = str(time.mktime(date.timetuple()) * 1000)
+            current_time = str(int(time.mktime(date.timetuple()) * 1000))
         print(current_time)
